Display a favorite add/delete button on the single app page
