+++ frontend/src/components/AppGrid/FavoriteAppGrid/FavAppGrid.css
@@ -28,7 +28,3 @@
 .FavAppGrid-Image:hover {
   cursor: pointer;
 }
-
-.FavAppGrid-FavIcon:hover {
-  cursor: pointer;
-}
