+++ frontend/src/components/AppGrid/FavoriteAppGrid/FavAppGrid.js
@@ -9,9 +9,7 @@ import {
   CardFooter,
 } from "reactstrap"
 import "./FavAppGrid.css"
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome"
-import { faHeart } from "@fortawesome/free-solid-svg-icons"
-import updateMyFavoriteApp from "../../../shared/fetchActions/updateMyFavoriteApp"
+import FavoriteButton from "./FavoriteButton"
 
 const FavAppGrid = ({ apps, appUrls }) => {
   return (
@@ -45,35 +43,8 @@ const RenderSingleApp = ({
         <CardText>{description}</CardText>
       </CardBody>
       <CardFooter>
-        <RenderFavoriteButton appId={id} isFavorite={isFavorite} />
+        <FavoriteButton appId={id} isFavorite={isFavorite} />
       </CardFooter>
     </Card>
-  )
-}
-
-const RenderFavoriteButton = ({ appId, isFavorite }) => {
-  const [isFav, setIsFavorite] = React.useState(isFavorite)
-
-  const toggleFavorite = () => {
-    // We set the favorite on the client side and asynchronously
-    // Ask the backend to also update the favorite
-    setIsFavorite(isFav => !isFav)
-
-    // TODO : Ideally the server update would be
-    // TODO : debounced in case a user is rapidly clicking the button, but perhaps that
-    // TODO : can be a later enhancement.
-    updateMyFavoriteApp(appId, !isFav)
-  }
-
-  return (
-    <FontAwesomeIcon
-      icon={faHeart}
-      color={isFav ? "red" : "grey"}
-      onClick={event => {
-        event.preventDefault()
-        toggleFavorite()
-      }}
-      className={"FavAppGrid-FavIcon"}
-    ></FontAwesomeIcon>
   )
 }
