+++ frontend/src/components/AppGrid/FavoriteAppGrid/FavoriteButton.css
@@ -1,0 +1,3 @@
+.FavButton-FavIcon:hover {
+  cursor: pointer;
+}
