+++ frontend/src/pages/apps/single-app.js
@@ -6,6 +6,7 @@ import ShareSocial from "../../components/share-social"
 import PaginatedComments from "../../components/Comments/PaginatedComments"
 import * as forFrontend from "../../shared/convertForFrontend"
 import getSingleApp from "../../shared/fetchActions/getSingleApp"
+import FavoriteButton from "../../components/AppGrid/FavoriteAppGrid/FavoriteButton"
 
 export const SingleApp = ({ appId }) => {
   const [app, setApp] = React.useState(null)
@@ -19,7 +20,7 @@ export const SingleApp = ({ appId }) => {
   }, [appId])
 
   if (app) {
-    const { name, description } = app
+    const { name, description, id, isFavorite } = app
     return (
       <div className="AnAppPage-container">
         <AppCarousel items={[app]} />
@@ -28,6 +29,7 @@ export const SingleApp = ({ appId }) => {
             <h1 className="display-3">{name}</h1>
             <p className="lead">{description}</p>
             <ShareSocial></ShareSocial>
+            <FavoriteButton appId={id} isFavorite={isFavorite}></FavoriteButton>
           </Container>
         </Jumbotron>
         <PaginatedComments comments={exampleComments}></PaginatedComments>
