+++ frontend/src/shared/convertForFrontend.js
@@ -6,6 +6,7 @@ export const convertApp = ({
   img,
   app_url,
   github_url,
+  is_favorite,
   id,
 }) => ({
   id,
@@ -14,6 +15,7 @@ export const convertApp = ({
   image: img,
   appUrl: app_url,
   githubUrl: github_url,
+  isFavorite: is_favorite,
 
   url: `/apps/${id}`,
   manageUrl: `/manage-apps/${id}/edit`,
